Add Docker Compose support for talkpush.api service
Added a new Docker Compose project to the solution file `CIBI_Webhooks.sln`, including project definition and configuration for build and release. Updated the `Dockerfile` to include a new `COPY` command for the `BuildingBlocks` project. Added a new property `DockerComposeProjectPath` to the `Talkpush.api.csproj` file. Created a `docker-compose.dcproj` file to define the Docker Compose project. Added `docker-compose.override.yml` and `docker-compose.yml` files to configure the `talkpush.api` service for development and define the service, respectively. Added a `launchSettings.json` file for debugging configuration. Created a `Dockerfile.original` for customized Dockerfile with multiple stages for building, publishing, and running the service.